+++ extensions/toggl-track/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Toggl Track Changelog
 
+## [New Feature] - 2025-02-05
+
+- Add ability to update time entry start and stop date
+
 ## [New Feature] - 2025-12-12
 
 - Add shortcut to "Create Similar Time Entry"